rename to usb-camera
