+++ application/ui/src/features/inference/sources/edit-source-form.component.tsx
@@ -7,10 +7,10 @@ import { ImageFolder } from './image-folder/image-folder.component';
 import { imagesFolderBodyFormatter } from './image-folder/utils';
 import { IpCamera } from './ip-camera/ip-camera.component';
 import { ipCameraBodyFormatter } from './ip-camera/utils';
+import { UsbCamera } from './usb-camera/usb-camera-fields.component';
+import { usbCameraBodyFormatter } from './usb-camera/utils';
 import { videoFileBodyFormatter } from './video-file/utils';
 import { VideoFile } from './video-file/video-file.component';
-import { usbCameraBodyFormatter } from './webcam/utils';
-import { WebcamFields } from './webcam/webcam-fields.component';
 
 interface EditSourceFormProps {
     config: SourceConfig;
@@ -25,7 +25,7 @@ export const EditSourceForm = ({ config, onSaved, onBackToList }: EditSourceForm
                 onSaved={onSaved}
                 config={config}
                 onBackToList={onBackToList}
-                componentFields={(state) => <WebcamFields defaultState={state} />}
+                componentFields={(state) => <UsbCamera defaultState={state} />}
                 bodyFormatter={usbCameraBodyFormatter}
             />
         );
