+++ application/ui/src/features/inference/sources/ip-camera/ip-camera.component.tsx
@@ -13,8 +13,14 @@ export const IpCamera = ({ defaultState }: IpCameraProps) => {
     return (
         <Flex direction='column' gap='size-200'>
             <TextField isHidden label='id' name='id' defaultValue={defaultState?.id} />
-            <TextField width={'100%'} label='Name' name='name' defaultValue={defaultState?.name} />
-            <TextField width={'100%'} label='Stream Url:' name='stream_url' defaultValue={defaultState?.stream_url} />
+            <TextField isRequired width={'100%'} label='Name' name='name' defaultValue={defaultState?.name} />
+            <TextField
+                isRequired
+                width={'100%'}
+                label='Stream Url:'
+                name='stream_url'
+                defaultValue={defaultState?.stream_url}
+            />
             <Switch
                 name='auth_required'
                 aria-label='Require Authentication'
