+++ application/ui/src/features/inference/sources/source-options.tsx
@@ -20,10 +20,10 @@ import { ImageFolder } from './image-folder/image-folder.component';
 import { getImagesFolderInitialConfig, imagesFolderBodyFormatter } from './image-folder/utils';
 import { IpCamera } from './ip-camera/ip-camera.component';
 import { getIpCameraInitialConfig, ipCameraBodyFormatter } from './ip-camera/utils';
+import { UsbCamera } from './usb-camera/usb-camera-fields.component';
+import { getUsbCameraInitialConfig, usbCameraBodyFormatter } from './usb-camera/utils';
 import { getVideoFileInitialConfig, videoFileBodyFormatter } from './video-file/utils';
 import { VideoFile } from './video-file/video-file.component';
-import { getUsbCameraInitialConfig, usbCameraBodyFormatter } from './webcam/utils';
-import { WebcamFields } from './webcam/webcam-fields.component';
 
 interface SourceOptionsProps {
     onSaved: () => void;
@@ -40,16 +40,14 @@ export const SourceOptions = ({ onSaved, hasHeader, children }: SourceOptionsPro
                 defaultActiveInput={null}
                 items={[
                     {
-                        label: 'Webcam',
+                        label: 'Usb Camera',
                         value: 'usb_camera',
                         icon: <WebcamIcon width={'24px'} />,
                         content: (
                             <AddSource
                                 onSaved={onSaved}
                                 config={getUsbCameraInitialConfig()}
-                                componentFields={(state: USBCameraSourceConfig) => (
-                                    <WebcamFields defaultState={state} />
-                                )}
+                                componentFields={(state: USBCameraSourceConfig) => <UsbCamera defaultState={state} />}
                                 bodyFormatter={usbCameraBodyFormatter}
                             />
                         ),
