+++ application/ui/src/features/inference/sources/usb-camera/utils.ts
